+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
-## 4.3.4
+## [4.3.5]
+
+- rename add_pin_square to add_pin_rectangle
+- add_pin_path to gf.add_pins
+
+## [4.3.4](https://github.com/gdsfactory/gdsfactory/pull/215)
 
 - tidy3d improvements:
     - get_simulation and write_sparameters accepts componentOrFactory